+++ framework/src/FluentSocket/Codecs/RequestMessage.cs
@@ -5,14 +5,14 @@ namespace FluentSocket.Codecs
 {
     public class RequestMessage : Message
     {
-        public int Code { get; set; }
+        public short Code { get; set; }
 
         public RequestMessage()
         {
 
         }
 
-        public RequestMessage(int code, byte[] body) : base(ObjectId.GenerateNewStringId(), MessageType.Request, body, DateTime.Now, null)
+        public RequestMessage(short code, byte[] body) : base(ObjectId.GenerateNewStringId(), MessageType.Request, body, DateTime.Now, null)
         {
             Code = code;
         }
